+++ tests/Commands/ReformatCommandTest.php
@@ -76,6 +76,56 @@ CONTENT;
     }
 
     #[Test]
+    public function sortsExportIgnoresFromDirectoriesToFiles(): void
+    {
+        $analyser = (new Analyser(new Finder(new PhpPreset())))->setDirectory($this->temporaryDirectory);
+        $repository = new GitattributesFileRepository($analyser);
+        $command = new ReformatCommand($analyser, $repository);
+
+        $gitattributesContent = <<<CONTENT
+/.gitattributes export-ignore
+/.gitignore export-ignore
+/CHANGELOG.md export-ignore
+/example export-ignore
+/README.rst export-ignore
+/tests export-ignore
+/.editorconfig export-ignore
+/.github export-ignore
+/contributing.rst export-ignore
+CONTENT;
+
+        $this->createTemporaryGitattributesFile($gitattributesContent);
+
+        $this->createTemporaryFiles(['.gitignore'], ['example', 'tests', '.github']);
+
+        $result = TestCommand::for($command)
+            ->addArgument($this->temporaryDirectory)
+            ->addOption('dry-run')
+            ->addOption('sort-from-directories-to-files')
+            ->execute()
+            ->assertSuccessful();
+
+        $expectedGitattributesContent = <<<CONTENT
+.github/          export-ignore
+example/          export-ignore
+tests/            export-ignore
+.editorconfig     export-ignore
+.gitattributes    export-ignore
+.gitignore        export-ignore
+CHANGELOG.md      export-ignore
+contributing.rst  export-ignore
+README.rst        export-ignore
+CONTENT;
+
+        $this->assertSame(\trim($expectedGitattributesContent), \trim($result->output()));
+
+        $this->assertStringEqualsFile(
+            $this->temporaryDirectory . DIRECTORY_SEPARATOR . '.gitattributes',
+            $gitattributesContent
+        );
+    }
+
+    #[Test]
     public function addASlashToExportIgnoredDirectories(): void
     {
         $analyser = (new Analyser(new Finder(new PhpPreset())))->setDirectory($this->temporaryDirectory);